+++ packages/iml-query/tests/test_query.py
@@ -3,10 +3,10 @@ from inline_snapshot import snapshot
 from iml_query.queries import (
     DECOMP_QUERY_SRC,
     INSTANCE_QUERY_SRC,
-    QCHECK_QUERY_SRC,
+    TEST_QUERY_SRC,
     VERIFY_QUERY_SRC,
     InstanceCapture,
-    QCheckCapture,
+    TestCapture,
     VerifyCapture,
 )
 from iml_query.tree_sitter_utils import (
@@ -141,7 +141,7 @@ instance (fun x -> x > 0) [@@by auto]
     assert instance_capture.instance_attr.text == b'[@@by auto]'
 
 
-def test_qcheck_parsing():
+def test_test_parsing():
     iml = """\
 let f = fun x -> x + 1
 
@@ -149,13 +149,27 @@ let g = fun x -> x + 2
 
 let f_g_x_gt_3 = fun x -> f (g x) - x > 3
 
-qcheck f_g_x_gt_3
+test f_g_x_gt_3
 """
     parser = get_parser()
     tree = parser.parse(bytes(iml, encoding='utf8'))
 
-    qc_matches = run_query(mk_query(QCHECK_QUERY_SRC), node=tree.root_node)
-    assert len(qc_matches) == 1
-    qc_capture = QCheckCapture.from_ts_capture(qc_matches[0][1])
-    assert qc_capture.qcheck_statement.text == b'qcheck f_g_x_gt_3'
-    assert qc_capture.qcheck_expr.text == b'f_g_x_gt_3'
+    t_matches = run_query(mk_query(TEST_QUERY_SRC), node=tree.root_node)
+    assert len(t_matches) == 1
+    t_capture = TestCapture.from_ts_capture(t_matches[0][1])
+    assert t_capture.test_statement.text == b'test f_g_x_gt_3'
+    assert t_capture.test_expr.text == b'f_g_x_gt_3'
+    assert t_capture.test_attr is None
+
+
+def test_test_parsing_with_attr():
+    iml = 'test (fun x -> x > 0) [@@by foo]\n'
+    parser = get_parser()
+    tree = parser.parse(bytes(iml, encoding='utf8'))
+
+    t_matches = run_query(mk_query(TEST_QUERY_SRC), node=tree.root_node)
+    assert len(t_matches) == 1
+    t_capture = TestCapture.from_ts_capture(t_matches[0][1])
+    assert t_capture.test_expr.text == b'(fun x -> x > 0)'
+    assert t_capture.test_attr is not None
+    assert t_capture.test_attr.text == b'[@@by foo]'