+++ packages/imandrax-tools/CHANGELOG.md
@@ -4,6 +4,9 @@ Versioning scheme: <IMANDRAX_API_VERSION>.<MINOR>.<PATCH>
 
 ## [Unreleased]
 
+## [20.0.0] - 2026-06-19
+- bump imandrax-api to v0.20
+
 ## [19.0.3] - 2026-04-25
 - improve diagnostic formatting: add severity error
 - add rule: opaque-attr-on-type-decl
@@ -19,4 +22,3 @@ Versioning scheme: <IMANDRAX_API_VERSION>.<MINOR>.<PATCH>
 ## [19.0.0] - 2026-04-23
 
 - add iml-eval-corpus subpackages
-